# Conflicts:
#	.github/workflows/tests-minimal.yml
#	tests/cypress/support/sql/database_6.0.0.sql

+++ system/ee/installer/updates/ud_6_00_00_b_1.php
@@ -1075,23 +1075,25 @@ class Updater
 
     private function migrateRte()
     {
-        ee()->smartforge->drop_table('rte_toolsets');
-        ee()->smartforge->drop_table('rte_tools');
-        ee()->db->data_cache = [];
+        if (ee()->db->table_exists('rte_toolsets')) {
+            ee()->smartforge->drop_table('rte_toolsets');
+            ee()->smartforge->drop_table('rte_tools');
+            ee()->db->data_cache = [];
 
-        require_once PATH_ADDONS . 'rte/upd.rte.php';
-        $Rte_upd = new \Rte_upd();
-        $Rte_upd->install_rte_toolsets_table();
+            require_once PATH_ADDONS . 'rte/upd.rte.php';
+            $Rte_upd = new \Rte_upd();
+            $Rte_upd->install_rte_toolsets_table();
 
-        $row_data = array(
-            'class' => 'Rte',
-            'method' => 'pages_autocomplete'
-        );
-        ee()->db->insert('actions', $row_data);
+            $row_data = array(
+                'class' => 'Rte',
+                'method' => 'pages_autocomplete'
+            );
+            ee()->db->insert('actions', $row_data);
 
-        ee()->db->where('name', 'Rte')->update('fieldtypes', ['version' => '2.0.0']);
+            ee()->db->where('name', 'Rte')->update('fieldtypes', ['version' => '2.0.0']);
 
-        ee()->db->where('module_name', 'Rte')->update('modules', ['module_version' => '2.0.0']);
+            ee()->db->where('module_name', 'Rte')->update('modules', ['module_version' => '2.0.0']);
+        }
     }
 
     private function addStickyChannelPreference()
